Rename mathieu headers
